requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
pyYAML
ensure
python-box
dvc
mlflow
dagshub